✨링크 저장 UI에 바로 반영도도록 변경

+++ src/domain/link/ILinkRepository.ts
@@ -1,4 +1,15 @@
+import { Link } from '@/types/links/link';
+
 export interface ILinkRepository {
+  checkAccess(linkId: string, userId: string): Promise<void>;
   deleteLink(linkId: string, userId: string): Promise<void>;
   createLink(data: { url: string; title: string; thumbnail: string; favicon: string; folderId: string; description: string; userId: string }): Promise<void>;
+  updateLink(
+    linkId: string,
+    data: {
+      title?: string;
+      description?: string;
+      isPin?: boolean;
+    }
+  ): Promise<Link>;
 }
